- refactoring:  * removed unnecessary type casts  * use try-with-resource where possible
git-svn-id: https://db-svn.informatik.uni-bremen.de/repos/stuff/use/trunk@5574 e95982d2-0d12-e879-a455-8ce5a0885042

+++ src/gui/org/tzi/use/gui/views/diagrams/behavior/DrawingUtil.java
@@ -141,7 +141,7 @@ public final class DrawingUtil {
 		g.transform(at);
 
 		// Draw horizontal arrow starting in (0, 0)
-		g.drawLine(0, 0, (int) len, 0);
+		g.drawLine(0, 0, len, 0);
 		g.fillPolygon(new int[] { len, len - ARR_SIZE, len - ARR_SIZE, len },
 				new int[] { 0, -ARR_SIZE, ARR_SIZE, 0 }, 4);
